Optimize acceptance test CI (#804)
* chore: split rpc.spec into smaller chunks


* fix: reset local-node files state after tests


* nit: resolve comment


* fix: error from merging


* chore: tweak test account balances


* chore: resolve conflicts


* chore: rename to batch


* nit: rename to batch


* chore: resolve conflicts


* chore: remove duplicated tests


---------
